Switch controller and views

+++ src/main/java/lv/freeradiusgui/dao/accountDAO/AccountDAO.java
@@ -12,7 +12,7 @@ public interface AccountDAO {
 
     boolean store(Account account);
 
-    Account getById(Long id);
+    Account getById(Integer id);
 
     Account getByLogin(String login);
